requirements.txt
Flask==2.3.2                  
mysql-connector-python==8.0.33  
numpy==1.23.5                 
scikit-learn==1.2.2           
